refactor(evals): split the results_extraction eval into modules under the limit
Copilot on #698: the eval module had reached 785 lines against the project's
300-line maximum, holding dataset loading, markdown parsing, bipartite matching,
eleven deterministic checks, judged scoring and viewer configuration at once.

Split along its seams, following the `verdict_checks.py` precedent for
suite-local helpers: `contract.py` (metric names and the vocabularies the ground
truth is written in), `parsing.py`, `matching.py`, `checks.py`, `criteria.py`,
`viewer.py`, with the task and its two scorers left in the eval module. Largest
is now 268 lines. Cross-module helpers lost their leading underscore; genuinely
private things kept theirs.

`matching.py` and `checks.py` are the two areas that have each produced a
reviewer-found bug on this branch, so having them separately readable and
separately tested is the point rather than a side effect.

The refactor is behaviour-neutral, and that is checked rather than asserted: the
eleven deterministic metrics are pure functions of (issues, expected,
document_lines), so recomputing them from the recorded baseline's stored outputs
reproduces all 330 values exactly.

Also adds RESULTS_EXTRACTION to the markdown delivery-contract suite, which the
migration should have done -- the skill-resolution parametrization had it, the
`report_issue` collection and markdown result contract did not.
