+++ evals_inspectai/e2e/results_extraction/criteria.py
@@ -1,0 +1,68 @@
+"""The judged criteria, and the delivery rendered for the grader to read.
+
+Criteria are stated as properties of the output, never as instructions to the
+grader, and `classification_grounded` carries the class taxonomy: without it the
+grader marked down rationales for correctly declining to treat an unrendered
+figure or a fixed-seed simulation as a gap.
+"""
+
+from evals_inspectai.common.simple_deep_agent_types import AgentCheckResult
+
+# Grader for the judged criteria, pinned per suite rather than inherited, so
+# changing this suite's judge does not silently re-grade any other.
+GRADER_MODEL = "openai/gpt-5.4"
+
+SHARED_CRITERIA: dict[str, str] = {
+    "classification_grounded": (
+        "Each result's rationale is grounded in what the document supplies: it names the "
+        "specific ingredients present or missing -- the data, parameters, equations, procedural "
+        "steps -- and how a reader would obtain anything missing, rather than asserting the "
+        "class or restating its definition. Where something is missing, it says what has to be "
+        "supplied without inventing the values.\n\n"
+        "Judge only the quality of that reasoning, not whether you would have picked the same "
+        "class. The four classes are defined as follows, and a rationale consistent with any of "
+        "them is grounded:\n"
+        "- Fully Reproducible: the logic is explained and everything needed is in the document, "
+        "appendices included.\n"
+        "- Reproducible with Web Search: the logic is explained, and what is missing is "
+        "individual published values a reader can look up from openly accessible sources.\n"
+        "- Reproducible with External Uploads: the logic is explained, and what is missing is a "
+        "bulk dataset open to any reader, which has to be downloaded.\n"
+        "- Not Reproducible: the logic is not explained, or what is missing cannot be obtained "
+        "at all -- confidential, proprietary, paid-access, withheld, or never recorded.\n\n"
+        "None of the following is a deficiency and a rationale must not be marked down for "
+        "declining to treat it as one: a figure or table not rendered in the text conversion; a "
+        "plot of the document's own model, which is regenerated by running that model; no "
+        "published code; values reported rounded; a reader having to fetch sources or re-run a "
+        "long computation; or a stochastic result whose seed and generator are fixed."
+    ),
+}
+
+
+def answer_text(result: AgentCheckResult) -> str:
+    """The delivery as prose for the grader: the inventory, then the report.
+
+    The issues are rendered rather than handed over as raw state JSON, because
+    what the criteria turn on -- which results were found, how each is
+    classified and defended -- is what a reader of the app sees, and JSON
+    punctuation would otherwise be most of the prompt.
+    """
+    entries = []
+    for issue in result.issues:
+        parts = [
+            f"### {issue.title}",
+            f"Severity: {issue.severity} | Lines {issue.start_line}-{issue.end_line}",
+            issue.description,
+        ]
+        if issue.long_description:
+            parts.append(issue.long_description)
+        if issue.suggested_action:
+            parts.append(f"Suggested action: {issue.suggested_action}")
+        entries.append("\n".join(parts))
+
+    inventory = "\n\n".join(entries) or "(no results reported)"
+    return (
+        f"=== RESULT INVENTORY ({len(result.issues)} reported) ===\n{inventory}\n\n"
+        f"=== REPORT ===\n{result.report_markdown}\n"
+    )
+
